+++ CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 - Fixed a small bug on the racket and ball colliders when the ball is too fast
 - Added a new collider on the walls to reduce the possibility of 'fall ball'
+- Now the version comes from the application